[device-minion-runtime] [SW-11723]: Updating payload size to 64-bit in firmware error event

+++ device-minion-runtime/src/MasterMinion/include/services/host_cmd_hdlr.h
@@ -49,6 +49,6 @@ int32_t Host_HP_Command_Handler(void *command_buffer, uint8_t sqw_hp_idx);
     \param payload (Optional) Payload.
     \return status success or negative error code
 */
-int32_t Device_Async_Error_Event_Handler(uint8_t error_type, uint32_t payload);
+int32_t Device_Async_Error_Event_Handler(uint8_t error_type, uint64_t payload);
 
 #endif /* HOST_CMD_HDLR_H */
